pyproject.toml
[project]
name = "mem"
version = "0.1.0"
description = "oh fuck it!"
authors = [
    {name = "jacobgoldenart", email = "golden.songs@gmail.com"},
]
dependencies = [
    "coqui-tts[server]>=0.22.1",
    "gruut>=2.3.4",
    "rich>=13.7.1",
    "pyaudio>=0.2.14",
    "watchdog>=4.0.0",
    "tavily-python>=0.3.3",
    "click>=8.1.7",
    "tomli>=2.0.1",
    "python-dotenv>=1.0.1",
    "uuid>=1.30",
    "openmeteo-requests>=1.2.0",
    "litellm[proxy]>=1.35.6",
    "langsmith>=0.1.48",
    "jinja2>=3.1.3",
    "tomli-w>=1.0.0",
    "ipapi>=1.0.4",
]
requires-python = "<3.12,>=3.9.0"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
